Combine aioutils.AioEmissionHoldLock and utils.EmissionHoldLock
