+++ pydispatch/utils.py
@@ -116,8 +116,12 @@ class InformativeWVDict(weakref.WeakValueDictionary):
     def _data_del_callback(self, key):
         self.del_callback(key)
 
-class EmissionHoldLock_(object):
+class EmissionHoldLock:
     """Context manager used for :meth:`pydispatch.dispatch.Dispatcher.emission_lock`
+
+    Supports use as a :term:`context manager` used in :keyword:`with` statements
+    and an :term:`asynchronous context manager` when used in
+    :keyword:`async with` statements.
 
     Args:
         event_instance: The :class:`~pydispatch.dispatch.Event` instance
@@ -135,6 +139,13 @@ class EmissionHoldLock_(object):
         self.event_instance = event_instance
         self.last_event = None
         self.held = False
+    @property
+    def aio_locks(self):
+        d = getattr(self, '_aio_locks', None)
+        if d is None:
+            d = self._aio_locks = {}
+        return d
+
     def acquire(self):
         if self.held:
             return
@@ -148,13 +159,35 @@ class EmissionHoldLock_(object):
             self.last_event = None
             self.held = False
             self.event_instance(*args, **kwargs)
+
+    async def acquire_async(self):
+        self.acquire()
+        lock = await self._build_aio_lock()
+        if not lock.locked():
+            await lock.acquire()
+    async def release_async(self):
+        lock = await self._build_aio_lock()
+        if lock.locked:
+            lock.release()
+        self.release()
+
     def __enter__(self):
         self.acquire()
         return self
     def __exit__(self, *args):
         self.release()
 
+    async def __aenter__(self):
+        await self.acquire_async()
+        return self
+    async def __aexit__(self, *args):
+        await self.release_async()
 
-from pydispatch.aioutils import AioEmissionHoldLock
-class EmissionHoldLock(EmissionHoldLock_, AioEmissionHoldLock):
-    pass
+    async def _build_aio_lock(self):
+        loop = asyncio.get_event_loop()
+        key = id(loop)
+        lock = self.aio_locks.get(key)
+        if lock is None:
+            lock = asyncio.Lock()
+            self.aio_locks[key] = lock
+        return lock
